career_kg/requirements.txt
neo4j>=5.0.0
neo4j-graphrag>=1.0.0
anthropic>=0.40.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
streamlit>=1.40.0
streamlit-agraph>=0.0.45
plotly>=5.18.0
cloudinary>=1.36.0
